==> code/03_finGap.R <==
# Global biodiversity finance flows and the nexus 
# Part of work for Chapter 6 of the IPBES Nexus Assessment
# Author: Andrea Pacheco

# this should be the script for visualizing the biodiversity financing gap

# libraries
library(dplyr)
library(tidyr)
library(tidyverse)
library(networkD3)
library(ggplot2)

# directory
wdmain <- "G:/My Drive/Projects/IPBES-Nexus/analyses/finFlows_nexus/"
